backend/requirements.txt
fastapi==0.109.2
mangum==0.17.0
aws-lambda-powertools==2.32.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0 